backend/requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.9.1
async-timeout==5.0.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
autopep8==2.2.0
billiard==4.2.1
black==23.11.0
boto3==1.34.0
botocore==1.34.162
celery==5.3.4
certifi==2025.7.14
cffi==1.17.1
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
cryptography==41.0.7
daphne==4.2.1
deprecation==2.1.0
Django==5.2.4
django-cors-headers==4.3.1
django-debug-toolbar==4.2.0
django-environ==0.12.0
django-extensions==4.1
django-filter==23.5
django-redis==5.4.0
django-silk==5.0.4
django-storages==1.14.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-spectacular==0.26.5
factory-boy==3.3.0
Faker==37.4.2
flake8==6.1.0
gotrue==2.12.3
gprof2dot==2025.4.14
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
inflection==0.5.1
iniconfig==2.1.0
isort==5.12.0
jmespath==1.0.1
jsonschema==4.24.1
jsonschema-specifications==2025.4.1
kombu==5.5.4
mccabe==0.7.0
msgpack==1.1.1
mypy_extensions==1.1.0
mysqlclient==2.2.7
packaging==25.0
pathspec==0.12.1
Pillow==10.1.0
platformdirs==4.3.8
pluggy==1.6.0
postgrest==1.1.1
prompt_toolkit==3.0.51
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.11.1
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pyflakes==3.1.0
PyJWT==2.10.1
PyMySQL==1.1.0
pyOpenSSL==25.1.0
pytest==7.4.3
pytest-django==4.7.0
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
realtime==2.6.0
redis==5.0.1
referencing==0.36.2
requests==2.31.0
rpds-py==0.26.0
s3transfer==0.9.0
sentry-sdk==1.38.0
service-identity==24.2.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
storage3==0.12.0
StrEnum==0.4.15
supabase==2.17.0
supafunc==0.10.1
Twisted==25.5.0
txaio==25.6.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
websockets==15.0.1
whitenoise==6.9.0
zope.interface==7.2
tensorflow==2.15.0
pandas==2.0.3
scikit-learn==1.3.2
